Add Gigya configuration
