+++ OwnID.Web.Gigya/Configuration/GigyaConfiguration.cs
@@ -1,14 +1,5 @@
-namespace OwnID.Web.Gigya
+namespace OwnID.Web.Gigya.Configuration
 {
-    public interface IGigyaConfiguration
-    {
-        string DataCenter { get; set; }
-        string ApiKey { get; set; }
-        string UserKey { get; set; }
-        string SecretKey { get; set; }
-        GigyaLoginType LoginType { get; set; }
-    }
-
     public class GigyaConfiguration : IGigyaConfiguration
     {
         public string DataCenter { get; set; }
